+++ geometry_grid.h
@@ -54,7 +54,7 @@ struct GridBox {
         uint64_t z;
     } size;
 
-    int32_t level;
+    uint32_t level;
 };
 
 #define box_create1(_px, _py, _pz, _level)               \
@@ -62,17 +62,17 @@ struct GridBox {
 #define box_createN(_px, _py, _pz, _sx, _sy, _sz, _level)               \
     (struct GridBox){ .position = { .x = _px, .y = _py, .z = _pz }, .size = { .x = _sx, .y = _sy, .z = _sz }, .level = _level };
 
-struct GridBox* grid_levelbox(struct Grid* grid, struct GridPages* pages, int32_t level, struct GridBox* box);
+struct GridBox* grid_levelbox(struct Grid* grid, struct GridPages* pages, uint32_t level, struct GridBox* box);
 
-struct GridBox* grid_pagebox_xyz(struct Grid* grid, struct GridPages* pages, uint64_t x, uint64_t y, uint64_t z, int32_t level, struct GridBox* box);
-struct GridBox* grid_pagebox(struct Grid* grid, struct GridPages* pages, uint64_t page, int32_t level, struct GridBox* box);
+struct GridBox* grid_pagebox_xyz(struct Grid* grid, struct GridPages* pages, uint64_t x, uint64_t y, uint64_t z, uint32_t level, struct GridBox* box);
+struct GridBox* grid_pagebox(struct Grid* grid, struct GridPages* pages, uint64_t page, uint32_t level, struct GridBox* box);
 
 struct GridBox* grid_levelup(struct Grid* grid, struct GridPages* pages, struct GridBox* box);
 struct GridBox* grid_leveldown(struct Grid* grid, struct GridPages* pages, struct GridBox* box);
 
 struct GridIndex {
     uint64_t page;
-    int32_t level;
+    uint32_t level;
     uint64_t cell;
 };
 struct GridIndex* grid_index_xyz(struct Grid* grid, struct GridPages* pages, struct GridBox* box, uint64_t x, uint64_t y, uint64_t z, struct GridIndex* index);
@@ -84,8 +84,8 @@ struct GridSize {
     uint64_t z;
     uint64_t array;
 };
-struct GridSize* grid_levelsize(struct Grid* grid, struct GridPages* pages, int32_t level, struct GridSize* size);
-struct GridSize* grid_pagesize(struct GridPages* pages, int32_t level, struct GridSize* size);
+struct GridSize* grid_levelsize(struct Grid* grid, struct GridPages* pages, uint64_t level, struct GridSize* size);
+struct GridSize* grid_pagesize(struct GridPages* pages, uint64_t level, struct GridSize* size);
 
 void grid_create(uint64_t x, uint64_t y, uint64_t z,
                  struct Grid* grid);
@@ -93,8 +93,8 @@ void grid_pages(struct Grid* grid, uint64_t x, uint64_t y, uint64_t z,
                 struct GridPages* pages);
 void grid_dump(struct Grid grid, struct GridPages pages);
 
-void grid_alloc(struct GridPages* pages, uint64_t page, int32_t level);
-void grid_free(struct GridPages* pages, uint64_t page, int32_t level);
+void grid_alloc(struct GridPages* pages, uint64_t page, uint32_t level);
+void grid_free(struct GridPages* pages, uint64_t page, uint32_t level);
 
 void grid_clear(struct Grid* grid, struct GridPages* pages, struct GridBox* box);
 
@@ -109,12 +109,12 @@ void grid_orN(struct Grid* grid, struct GridPages* pages, struct GridBox* box, C
 
 void grid_shift(struct Grid* grid, struct GridPages* pages, struct GridBox* box, int32_t shift);
 
-void grid_pagein(struct Grid* grid, struct GridPages* pages, uint64_t page, int32_t level, char** in);
-void grid_pageout(struct Grid* grid, struct GridPages* pages, uint64_t page, int32_t level, char** out);
+void grid_pagein(struct Grid* grid, struct GridPages* pages, uint64_t page, uint32_t level, char** in);
+void grid_pageout(struct Grid* grid, struct GridPages* pages, uint64_t page, uint32_t level, char** out);
 
 void world_grid_create(struct Grid* grid,
                        struct GridPages* pages,
-                       int32_t level,
+                       uint32_t level,
                        float width,
                        float height,
                        float depth,
@@ -122,7 +122,7 @@ void world_grid_create(struct Grid* grid,
                        struct VboMesh* mesh);
 void world_grid_update(struct Grid* grid,
                        struct GridPages* pages,
-                       int32_t level,
+                       uint32_t level,
                        uint64_t page,
                        struct Cube* cube,
                        struct VboMesh* mesh);
